extract aave quote planning, drop client accessor
